Move CardWrapper component to shared location
