+++ src/components/admin/Billing/index.tsx
@@ -1,18 +1,18 @@
 import { Divider, Grid, Typography } from '@mui/material';
 import { getInvoicesBetween } from 'api/billing';
 import { authenticatedRoutes } from 'app/routes';
-import CardWrapper from 'components/admin/Billing/CardWrapper';
 import DateRange from 'components/admin/Billing/DateRange';
 import PaymentMethods from 'components/admin/Billing/PaymentMethods';
 import PricingTierDetails from 'components/admin/Billing/PricingTierDetails';
 import TenantOptions from 'components/admin/Billing/TenantOptions';
 import AdminTabs from 'components/admin/Tabs';
-import UsageByMonthGraph from 'components/graphs/UsageByMonthGraph';
 import GraphLoadingState from 'components/graphs/states/Loading';
 import GraphStateWrapper from 'components/graphs/states/Wrapper';
+import UsageByMonthGraph from 'components/graphs/UsageByMonthGraph';
 import AlertBox from 'components/shared/AlertBox';
-import BillingLineItemsTable from 'components/tables/BillLineItems';
+import CardWrapper from 'components/shared/CardWrapper';
 import BillingHistoryTable from 'components/tables/Billing';
+import BillingLineItemsTable from 'components/tables/BillLineItems';
 import { endOfMonth, startOfMonth, subMonths } from 'date-fns';
 import usePageTitle from 'hooks/usePageTitle';
 import { useEffect, useMemo } from 'react';
@@ -21,13 +21,13 @@ import { FormattedMessage } from 'react-intl';
 import { useUnmount } from 'react-use';
 import { logRocketEvent } from 'services/shared';
 import { CustomEvents } from 'services/types';
-import { useBillingStore } from 'stores/Billing/Store';
 import { useBilling_selectedInvoice } from 'stores/Billing/hooks';
+import { useBillingStore } from 'stores/Billing/Store';
 import { useTenantStore } from 'stores/Tenant/Store';
 import useConstant from 'use-constant';
-import { TOTAL_CARD_HEIGHT, invoiceId } from 'utils/billing-utils';
-import { AdminBillingProps } from './types';
+import { invoiceId, TOTAL_CARD_HEIGHT } from 'utils/billing-utils';
 import BillingLoadError from './LoadError';
+import { AdminBillingProps } from './types';
 
 const routeTitle = authenticatedRoutes.admin.billing.title;
