+++ src/components/shared/Entity/Details/Overview/NotificationSettings/index.tsx
@@ -1,7 +1,7 @@
 import { Box, Stack, useMediaQuery, useTheme } from '@mui/material';
-import CardWrapper from 'components/admin/Billing/CardWrapper';
 import MessageWithButton from 'components/content/MessageWithButton';
 import AlertBox from 'components/shared/AlertBox';
+import CardWrapper from 'components/shared/CardWrapper';
 import DataProcessingSetting from 'components/shared/Entity/Details/Overview/NotificationSettings/DataProcessingSetting';
 import Error from 'components/shared/Error';
 import { ErrorDetails } from 'components/shared/Error/types';
